Add recipe for GB and brownies, add links to index
Added headers in index to specify what kind of food
it is making it easier for the user to navigate.
Added a new recipe for garlic bread in garlic-bread.html
It consists of an img, the ingredients and the steps.
Added a return to homepage button to both recipes to allow
users to navigate back to the homepage in index.html.