Add new css file and link to all html file
Added a new css file in order to add design to the
homepage and other recipe pages. Design additions
are simple such as changing the background colour,
adding borders, changing font,text colour, etc.
This addition to the project is only meant as
practice for the css I have learnt.